+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - Expand folder item by double-clicking a folder row in the file browser.
 - Directly open plist files from the file browser instead of previewing.
+- Improve alias files handling in the file browser.
 
 
 ### Fixes