+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Fixes
 
+- Fix an issue that the application crashed when an editor is opened while the Font window appears.
 - Fix an issue that document window size was not properly saved.